Move Metadata.length_score to matching module
