+++ test/test_metadata.py
@@ -302,26 +302,6 @@ class CommonTests:
             self.assertEqual(0, len(metadata.rawitems()))
             self.assertEqual(self.metadata.keys(), metadata.deleted_tags)
 
-        def test_length_score(self):
-            results = (
-                (20000, 0, 0.333333333333),
-                (20000, 10000, 0.666666666667),
-                (20000, 20000, 1.0),
-                (20000, 30000, 0.666666666667),
-                (20000, 40000, 0.333333333333),
-                (20000, 50000, 0.0),
-                (20000, None, 0.0),
-                (None, 2000, 0.0),
-                (None, None, 0.0),
-            )
-            for a, b, expected in results:
-                actual = Metadata.length_score(a, b)
-                self.assertAlmostEqual(
-                    expected,
-                    actual,
-                    msg="a={a}, b={b}".format(a=a, b=b),
-                )
-
         def test_compare_is_equal(self):
             m1 = Metadata()
             m1["title"] = "title1"
